// src/models/search.ts
export interface LiteratureType {
    name: string;
    value: number;
}

export interface IExampleData {
    entityId: string;
    entityName: string;
    entityType: string;
    orgId: string;
    orgName: string;
}
export interface ISearchParam {
    entityId: string;
    entityName: string;
    entityType: string;
}